修复路由重复及头像的bug

+++ src/components/Trash.vue
@@ -9,8 +9,8 @@
       <ul class="notes">
         <li v-for="note in trashNotes">
           <router-link :to="`/trash?noteId=${note.id}`">
-            <span class="date">{{note.newUpdatedAt}}</span>
-            <span class="title">{{note.title}}</span>
+            <span class="date">{{ note.newUpdatedAt }}</span>
+            <span class="title">{{ note.title }}</span>
           </router-link>
         </li>
       </ul>
@@ -19,17 +19,17 @@
     <div class="note-detail-wrapper">
       <div class="note-detail">
         <div class="note-bar" v-if="true">
-          <span> 创建日期: {{curTrashNote.newCreatedAt}}</span>
+          <span> 创建日期: {{ curTrashNote.newCreatedAt }}</span>
           <span> | </span>
-          <span> 更新日期: {{curTrashNote.newUpdatedAt}}</span>
+          <span> 更新日期: {{ curTrashNote.newUpdatedAt }}</span>
           <span> | </span>
-          <span> 所属笔记本: {{belongTo}}</span>
+          <span> 所属笔记本: {{ belongTo }}</span>
 
           <a class="btn action" @click="onRevert">恢复</a>
           <a class="btn action" @click="onDelete">彻底删除</a>
         </div>
         <div class="note-title">
-          <span>{{curTrashNote.title}}</span>
+          <span>{{ curTrashNote.title }}</span>
         </div>
         <div class="editor">
           <div class="preview markdown-body" v-html="compiledMarkdown"></div>
@@ -40,10 +40,10 @@
 </template>
 
 <script>
-import MarkdownIt from 'markdown-it'
-import { mapGetters, mapMutations, mapActions } from 'vuex'
+import MarkdownIt from 'markdown-it';
+import {mapGetters, mapMutations, mapActions} from 'vuex';
 
-let md = new MarkdownIt()
+let md = new MarkdownIt();
 
 export default {
   computed: {
@@ -53,8 +53,8 @@ export default {
       'belongTo'
     ]),
 
-    compiledMarkdown () {
-      return md.render(this.curTrashNote.content||'')
+    compiledMarkdown() {
+      return md.render(this.curTrashNote.content || '');
     }
   },
 
@@ -72,28 +72,57 @@ export default {
     ]),
 
     onDelete() {
-      this.deleteTrashNote({ noteId: this.curTrashNote.id })
+      this.$confirm('删除后将无法恢复', '确定删除？', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      })
+        .then(() => {
+          this.deleteTrashNote({noteId: this.curTrashNote.id});
+        })
+        .then(() => {
+          this.setCurTrashNote();
+          this.$router.replace({
+            path: '/trash',
+            query: {noteId: this.curTrashNote.id}
+          });
+        });
+
     },
 
     onRevert() {
-      this.revertTrashNote({ noteId: this.curTrashNote.id })
+      this.revertTrashNote({noteId: this.curTrashNote.id}).then(() => {
+        this.setCurTrashNote();
+        this.$router.replace({
+          path: '/trash',
+          query: {
+            noteId: this.curTrashNote.id,
+          }
+        });
+      });
     }
   },
 
   created() {
-    this.checkLogin({ path: '/login' })
-    this.getNotebooks()
+    this.checkLogin({path: '/login'});
+    this.getNotebooks();
     this.getTrashNotes()
       .then(() => {
-        this.setCurTrashNote({ curTrashNoteId: this.$route.query.noteId })
-      })
+        this.setCurTrashNote({curTrashNoteId: this.$route.query.noteId});
+        this.$router.replace({
+          path: '/trash',
+          query: {
+            noteId: this.curTrashNote.id,
+          }
+        });
+      });
   },
 
-  beforeRouteUpdate (to, from, next) {
-    this.setCurTrashNote({ curTrashNoteId: to.query.noteId})
-    next()
+  beforeRouteUpdate(to, from, next) {
+    this.setCurTrashNote({curTrashNoteId: to.query.noteId});
+    next();
   }
-}
+};
 </script>
 
 <style lang="less">
